+++ src/domible/elements/__init__.py
@@ -9,7 +9,7 @@ from domible.elements.baseElements import BaseElement, BaseVoidElement
 
 from domible.elements.roots import Html
 from domible.elements.meta import Base, Head, Meta, Style, Title 
-from domible.elements.inline import Anchor, Div, Paragraph, Span
+from domible.elements.inline import Anchor, Div, Paragraph, Span, Summary, Details
 from domible.elements.lists import DescriptionTerm, DescriptionDef, DListItem, ListItem, OrderedList, UnorderedList, MenuList, DescriptionList
 from domible.elements.scripting import Canvas, NoScript, Script, Template 
 from domible.elements.sectioning import Article, Body, Footer, Header, Heading, HorizontalRule, Main, Nav, Section