+++ src/domible/elements/inline.py
@@ -35,4 +35,32 @@ class Span(BaseElement):
         super().__init__(tag="span", contents=contents, **kwArgs)
 
 
+class Summary(BaseElement):
+    """create a Summary element
+    used within a details element to provide a label for a collapseable widget.
+    For accessibility, requiring some content be provided
+    """
+
+    def __init__(self, contents: Any, **kwArgs):
+        if not contents:
+            raise ValueError("Summary element must have some contents")
+        super().__init__(tag="summary", contents=contents, **kwArgs)
+
+
+class Details(BaseElement):
+    """create a details element
+    used to create a collapseable widget.
+    For accessibility, requiring a summary element be provided
+    """
+
+    def __init__(self, summary: Summary,  contents: Any = None, **kwArgs):
+        if not summary or not isinstance(summary, Summary):
+            raise ValueError("Details element must contain a Summary element")
+        if isinstance(contents, list): 
+            contents = contents.insert(0, summary)
+        else:
+            contents  = [summary] + [contents] if contents else summary
+        super().__init__(tag="details", contents=contents, **kwArgs)
+
+
 ## end of file